requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Science & Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
Pillow>=9.5.0

# Progress Bars & Logging
tqdm>=4.65.0
loguru>=0.7.0

# Development Tools
jupyter>=1.0.0
ipython>=8.12.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
mkdocs>=1.4.0

# Utilities
python-dotenv>=1.0.0
typer>=0.9.0
requests>=2.31.0
celluloid>=0.2.0  # For animations

# Build Tools
flit_core>=3.2.0,<4.0.0